FEAT: deposit improvements and features related to bank slip
The deposit of funds, implemented using the issuance of root asset, started to support that any user associated with the domain can command the action. Although it does not make much sense, it was adopted this way because the Conductor has not yet made available webhooks for deposit notification into account.
Also implemented endpoints that expose the functions of registering and consulting a bank slip.