FEAT: improvements in the creation and search of bank splip
In order to avoid that limitations in integration with Conductor make it impossible to complete the implementation of the feature, the application started to generate the barcode, whose data (digitable line and the barcode) are saved in the local database. It was also implemented an improvement in the search of the bank slip that allows the search using as identifiers the id of the resource, the barcode as well as the digitable line.